Move some functions from lxc driver to execdriver utils
Some of the internal functions of the LXC driver would be needed for a
Libvirt driver as those two have quite some similar constraints. Moving
them to execdriver/utils.go to help code sharing.
